clean architecture
