--- busloc/ContentView.swift
@@ -1,16 +1,0 @@
-import SwiftUI
-
-struct ContentView: View {
-    @Environment(\.modelContext) private var modelContext
-
-    var body: some View {
-        NavigationStack{
-            HomeView()
-        }
-    }
-}
-
-#Preview {
-    ContentView()
-        .modelContainer(for: Item.self, inMemory: true)
-}
