--- busloc/Data/BusData.swift
@@ -1,52 +1,0 @@
-final class BusData {
-    static func getData() -> [Bus] {
-        return [
-            .init(name: "#1", route: [
-                "Intermoda", "Cosmo", "Verdant View", "Eternity", "Simplicity 2", "Edutown 1", "Edutown 2",
-                "ICE 1", "ICE 2", "ICE Business Park", "ICE 6", "ICE 5", "GOP 1", "SML Plaza", "The Breeze",
-                "CBD Timur 1", "CBD Timur 2", "Nava Park 1", "SWA 2", "Giant", "Eka Hospital 1", "Puspitaloka",
-                "Polsek Serpong", "Ruko Madrid", "Pasar Modern Timur", "Griya Loka 1", "Sektor 1.3", "Griya Loka 2",
-                "Santa Ursula 1", "Santa Ursula 2", "Sentra Onderdil", "Autoparts", "Eka Hospital 2", "East Business District",
-                "SWA 1", "Greencove", "The Breeze", "CBD Timur 1", "CBD Selatan 1", "AEON Mall 1", "CBD Barat 2",
-                "Simpilicity 1", "Cosmo", "Verdant View", "Intermoda", "Icon Business Park", "Masjid Al-Ukhuwah",
-                "Divena & Deshna", "Avani", "Amarilla", "Chadnya", "Atmajaya", "Intermoda"
-            ], etaStart: "10:30", etaEnd: "11:30"),
-            .init(name: "#2", route: [
-                "Greenwich Park Office", "Jadeite", "De Maja", "De Heliconia 2", "De Nara", "De Park 2", "Nava Park 2",
-                "Giardina", "Collinare", "Foglio", "Studento 2", "Albera", "Foresta 1", "GOP 1", "SML Plaza", "The Breeze",
-                "CBD Timur 1", "CBD Timur 2", "Nava Park 1", "SWA 2", "Giant", "Eka Hospital 1", "Puspitaloka",
-                "Polsek Serpong", "Ruko Madrid", "Pasar Modern Timur", "Griya Loka 1", "Sektor 1.3", "Griya Loka 2",
-                "Santa Ursula 1", "Santa Ursula 2", "Sentra Onderdil", "Autoparts", "Eka Hospital 2", "East Business District",
-                "SWA 1", "Greencove", "The Breeze", "CBD Timur 1", "CBD Timur 2", "Simpang Foresta", "Allevare", "Fiore",
-                "Studento 1", "Naturale", "Fresco", "Primavera", "Foresta 2", "FBL 5", "Courts Mega Store", "Q BIG 1",
-                "Lulu", "Greenwich Park 1"
-            ], etaStart: "11:00", etaEnd: "11.30"),
-            .init(name: "#3", route: [
-                "Intermoda", "Simplicity 2", "Edutown 1", "Edutown 2", "ICE 1", "ICE 2", "ICE Business Park", "ICE 6", "ICE 5",
-                "EPICON", "Gramedia", "Astra", "Courts Mega Store", "Q BIG 1", "Lulu", "Greenwich Park 1", "Jadeite",
-                "De Maja", "De Heliconia 2", "De Nara", "De Park 2", "Nava Park 2", "GOP 1", "Giardina", "Collinare", "Foglio",
-                "Studento 2", "Albera", "Foresta 1", "GOP 1", "SML Plaza", "The Breeze", "Casa De Parco 2", "B M C",
-                "Digital Hub 1", "Digital Hub 2", "Verdant View", "Eternity", "Intermoda"
-            ], etaStart: "11:30", etaEnd: "12:30"),
-            .init(name: "#4", route: [
-                "Intermoda", "Icon Centro", "Horizon Broadway", "Extreme Park", "Saveria", "Casa De Parco 1", "SML Plaza",
-                "The Breeze", "CBD Timur 1", "CBD Selatan 1", "AEON Mall 1", "CBD Timur 2", "Simpang Foresta", "Allevare",
-                "Fiore", "Studento 1", "Naturale", "Fresco", "Primavera", "Foresta 2", "De Park 1", "De Frangpani",
-                "De Heliconia 1", "De Brassia", "Jadeite", "Greenwich Park 2", "Q Big 2", "Q Big 3", "BCA", "FBL 2", "FBL 1",
-                "ICE 1", "ICE 2", "ICE Business Park", "ICE 6", "ICE 5", "CBD Barat 1", "CBD Barat 2", "Simplicity 1", "Intermoda"
-            ], etaStart: "12:30", etaEnd: "13:30"),
-            .init(name: "#5", route: [
-                "The Breeze", "CBD Timur 1", "CBD Selatan 1", "Lobby AEON", "AEON Mall 2", "CBD Utara 3", "ICE 1", "ICE 2",
-                "ICE Business Park", "ICE 6", "ICE 5", "CBD Barat 1", "CBD Barat 2", "Lobby AEON", "AEON Mall 2",
-                "CBD Timur 2", "Nava Park 1", "Greencove", "The Breeze"
-            ], etaStart: "13:30", etaEnd: "14:30"),
-            .init(name: "#7", route: [
-                "Intermoda", "Simplicity 2", "Edutown 1", "Edutown 2", "ICE 1", "ICE 2", "ICE Business Park", "ICE 6", "ICE 5",
-                "EPICON", "Gramedia", "Astra", "Courts Mega Store", "Q BIG 1", "Lulu", "Q Big 2", "Q Big 3", "BCA", "FBL 2",
-                "FBL 1", "GOP 1", "SML Plaza", "The Breeze", "Casa De Parco 2", "B M C", "Digital Hub 1", "Saveria",
-                "Casa De Parco 1", "CBD Timur 1", "CBD Selatan 1", "Lobby AEON", "AEON Mall 1", "AEON Mall 2", "CBD Barat 2",
-                "Simplicity 1", "Intermoda"
-            ], etaStart: "14:30", etaEnd: "15:30"),
-            ]
-    }
-}
