+++ busloc/Domain/Models/Bus.swift
@@ -1,3 +1,10 @@
+//
+//  Bus.swift
+//  busloc
+//
+//  Created by Abim on 09/04/25.
+//
+
 struct Bus: Hashable {
     let name: String
     let route: [String]
