--- busloc/Presentation/ContentView.swift
@@ -1,15 +1,0 @@
-import SwiftUI
-
-struct ContentView: View {
-    @Environment(\.modelContext) private var modelContext
-
-    var body: some View {
-        NavigationStack{
-            HomeView()
-        }
-    }
-}
-
-#Preview {
-    ContentView()
-}
